implement tests and code for fuel decreasing while train moving (and turning off movement when fuel hits zero)

+++ BirkbeckProject-Game-MVP/Assets/HUD/HUDFuelScript.cs
@@ -12,11 +12,11 @@ public class HUDFuelScript : MonoBehaviour
     public TrainDataScriptableObject trainData;
     public TMP_Text textObject;
     private string textString;
-    public float currentFuel;
+    public int currentFuel;
 
     private void Update()
     {
-        currentFuel = trainData.GetTrainFuel();
+        currentFuel = (int)trainData.GetTrainFuel();
         textString = "Fuel: " + currentFuel.ToString();
         textObject.text = textString;
     }
